Mock Login Flow
This commit introduces a mocked login flow.

Key changes:
- In `LoginPage`, upon successful login (`LoginStatus.success`), the user is now navigated to the first `AppShellBranch`.
- The `LoginButton`'s `onPressed` logic is updated to trigger the `LoginSubmitButtonPressed` event in `LoginBloc` only if the email is "brocolisoftware@gmail.com" or "andrea.lopez@gmail.com".
- In `AuthRepository`, the actual API calls for `signUp` and `signIn` are commented out.
    - Instead of deleting and fetching a token, it now checks if a token exists, deletes it if it does, and then uses the provided email as the token.
    - This allows for a mocked login without actual backend interaction.

+++ lib/src/modules/auth/data/repositories_impl/auth_repository.dart
@@ -28,11 +28,15 @@ class AuthRepository implements IAuthRepository {
         name: '$_source.signUp()',
       );
 
-      await _storage.deleteValue(_storageTokenKey);
+      final value = await _storage.getValue(_storageTokenKey);
+      if (value != null) {
+        await _storage.deleteValue(_storageTokenKey);
+      }
 
-      final token = await _authApi.signUpUser(
-        params: SignUpUserParamsModel.fromEntity(params),
-      );
+      final token = params.email;
+      //await _authApi.signUpUser(
+      //  params: SignUpUserParamsModel.fromEntity(params),
+      //);
 
       await _storage.storeValue(key: _storageTokenKey, value: token);
 
@@ -59,11 +63,17 @@ class AuthRepository implements IAuthRepository {
         name: '$_source.signIn()',
       );
 
-      await _storage.deleteValue(_storageTokenKey);
+      final value = await _storage.getValue(_storageTokenKey);
+      if (value != null) {
+        await _storage.deleteValue(_storageTokenKey);
+      }
 
-      final token = await _authApi.signInUser(
-        params: SignInUserParamsModel.fromEntity(params),
-      );
+      final token = params.email;
+      //await _storage.deleteValue(_storageTokenKey);
+
+      //final token = await _authApi.signInUser(
+      //  params: SignInUserParamsModel.fromEntity(params),
+      //);
 
       await _storage.storeValue(key: _storageTokenKey, value: token);
